requirements.txt
PyQt6==6.6.0
spotipy==2.23.0
requests==2.31.0
google-generativeai==0.3.2
python-dotenv==1.0.0  # Recommended for managing environment variables